Fix semaphore init bug and work_queue items getting skipped
Semaphore was getting initialized to 1 instead of 0, which is just wrong.

Additionally, I changed to only using a Semaphore for the flow control
during ThreadMain, which just felt better.  Doing so required calling
AtomicCompareExchange in a loop so as to not call ThreadSleep multiple
times if the Exchange failed to go thorugh.  It also required changing
the win32 CreateSemaphore to cap out at LONG_MAX instead of the number of
threads.