Inflate volume of voxels initialized by one on each side of world_chunk
In preparation for generating a world chunk mesh in a multithreaded code
path I modified the world chunk initialization code to world on a
synthetic chunk which is inflated by one unit on each side.  This way we
will not have to deal with pulling neighboring chunks into the CPU cache
to build the complete mesh.