No longer modify the DeltaX returned by BufferChar based on the ClipResult
We'd rather have accurate DrawBounds computations even for stuff that's
been clipped off the window.  This is beneficial because if DrawBounds
computes a DeltaX of 0 in AdvanceClip we don't get a newline, so if
short lines get completely clipped the net effect is that line
effectively dissappears, which bumps lower lines up in the layout.