Fix editor preview jank
This is not a very good fix.  I moved the mesh generation onto the main
thread because I couldn't figure out the correct order-of-operations in
terms of storing the min corner offset (it's more fucky than you think).

This should all get moved to a thread at some point in the future.  The
fucky problem won't be a problem because then everything will be on the
thread (hopefully), and the offsets will just be obvious (I claim). GLHF